requirements.txt
Django==2.1
djangorestframework==3.9.4
gunicorn==19.9.0
nltk==3.4.1
pytz==2019.1
six==1.12.0
